+++ css-inline/src/html/node.rs
@@ -63,7 +63,7 @@ impl NodeId {
         NodeId::new(1)
     }
     #[inline]
-    pub(super) fn get(self) -> usize {
+    pub(crate) fn get(self) -> usize {
         self.0.get()
     }
 }